+++ R/utilis.R
@@ -8,14 +8,31 @@
 
 # Build data
 .buildData <- function(
-    data, ID, ts, sig, aLon, aLat, aType, aBearing, det_range) {
+    data, ID, ts, sig, aLon, aLat, aType, aBearing, aRange) {
+  # Check for missing columns
+  i <- !c(ID, ts, sig, aLon, aLat, aBearing) %in% names(data)
+  if (any(i)) {
+    stop(paste0(
+      "Missing columns: '",
+      paste(c(ID, ts, sig, aLon, aLat, aBearing)[i],
+        collapse = "', '"
+      ),
+      "'."
+    ))
+  }
   d <- data.frame(
     ID = data[[ID]],
     ts = data[[ts]],
     sig = data[[sig]],
     aLon = data[[aLon]],
     aLat = data[[aLat]],
-    aType = data[[aType]],
+    aType = if (is.list(aRange)) {
+      if (aType %in% names(data) && !is.null(aType)) {
+        data[[aType]]
+      } else {
+        stop("Missing 'aType' column.")
+      }
+    } else "any",
     aBearing = data[[aBearing]]
   )
   c <- !complete.cases(d)
@@ -28,20 +45,20 @@
   d <- arrange(d[!c, ], ID, ts)
 
   # Maximum detection range per antenna type
-  switch(typeof(det_range),
-    double = d$det_range <- det_range,
+  switch(typeof(aRange),
+    double = d$aRange <- aRange,
     list = {
-      if (all(d$aType %in% names(det_range))) {
-        d$det_range <- unlist(det_range[d$aType])
+      if (all(d$aType %in% names(aRange))) {
+        d$aRange <- unlist(aRange[d$aType])
       } else {
         stop(paste0(
           "Missing detection range for '",
-          paste(setdiff(unique(d$aType), names(det_range)), collapse = "', '"),
+          paste(setdiff(unique(d$aType), names(aRange)), collapse = "', '"),
           "'."
         ))
       }
     },
-    stop("'det_range' must be numeric or a named list.")
+    stop("'aRange' must be numeric or a named list.")
   )
   return(d)
 }