+++ tests/testthat/setup.R
@@ -1,5 +1,5 @@
 data(motusData)
-suppressMessages(loc <- locate(motusData, dtime = 10))
+suppressMessages(loc <- locate(motusData, dTime = 10))
 suppressMessages(loc$ID[1] <- 1)
 
 test_that("track warnings", {