+++ tests/testthat/test-locate.R
@@ -1,17 +1,30 @@
 test_that("locate error", {
   expect_error(
-    locate(motusData, det_range = "loo"),
-    "'det_range' must be numeric or a named list."
+    locate(motusData |> select(-sig)),
+    "Missing columns: 'sig'."
+    )
+  expect_error(
+    locate(motusData, aRange = "foo"),
+    "'aRange' must be numeric or a named list."
   )
   expect_error(
-    locate(motusData, det_range = list("yagi-6" = 12)),
+    locate(motusData, aType = "antType", aRange = list("yagi-6" = 12)),
     "Missing detection range for 'yagi-5'."
+  )
+  expect_error(
+    locate(motusData |> select(-antType),
+      aType = "antType", aRange = list("yagi-6" = 12)
+    ),
+    "Missing 'aType' column."
   )
 })
 
 test_that("locate result", {
   expect_message(
-    loc <- locate(motusData, det_range = list("yagi-5" = 10, "yagi-6" = 12)),
+    loc <- locate(motusData,
+      aType = "antType",
+      aRange = list("yagi-5" = 10, "yagi-6" = 12)
+    ),
     "Removed 10 detections containing missing values."
   )
   expect_true(is.data.frame(loc))